# Backend/backend/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-05-17T20:41:54+00:00

from __future__ import annotations

# modified by laia-gen-lib:

from pydantic import ConfigDict
from laiagenlib.Domain.LaiaBaseModel.LaiaBaseModel import LaiaBaseModel
from laiagenlib.Domain.LaiaUser.LaiaUser import LaiaUser
from laiagenlib.Domain.GeoJSON.Geometry import Type, Geometry, LineString, MultiLineString, MultiPoint, MultiPolygon, Point, Polygon

from datetime import datetime
from enum import Enum
from typing import List, Optional

from pydantic import BaseModel, Field, SecretStr


class Location(LaiaBaseModel):
    type: Type
    coordinates: List[float]


class Shop(LaiaBaseModel):
    name: str
    activity: Optional[str] = None
    category: str
    address: Optional[str] = None
    neighborhood: Optional[str] = None
    is_commercial_axis: Optional[bool] = None
    location: Location


class Transaction(LaiaBaseModel):
    user_id: str = Field(..., description='User ID', x_required=True, x_type="objectId")
    amount: float = Field(
        ..., description='Transaction amount (positive for credit, negative for debit)'
    )
    type: str = Field(
        ..., description='Transaction type (topup, send, receive, payment)'
    )
    description: str = Field(..., description='Transaction description', )
    counterparty: Optional[str] = Field(
        None, description='Email or ID of the other user'
    )
    created_at: Optional[datetime] = Field(
        None, description='Transaction date and time'
    )


class User(LaiaUser):
    model_config = ConfigDict(json_schema_extra={'x-auth': True})
    name: str = Field(..., description='User first name', )
    surnames: Optional[str] = Field(None, description='User last name', )
    email: str = Field(..., description='User email address', )
    password: SecretStr = Field(..., description='User password', )
    phone: Optional[str] = Field(None, description='User phone number', )


class Wallet(LaiaBaseModel):
    address: str = Field(..., description='The wallet address', x_required=True, x_unique=True)
    balance: Optional[float] = Field(0.0, description='The wallet balance', )
    currency: Optional[str] = Field('USD', description='The wallet currency', )
    user_id: str = Field(..., description='The user ID', x_required=True, x_type="objectId")