# Backend/openapi.yaml
components:
  schemas:
...
    Body_search_element_transactions__post:
      properties:
        filters:
          additionalProperties: true
          default: {}
          title: Filters
          type: object
        orders:
          additionalProperties: true
          default: {}
          title: Orders
          type: object
      title: Body_search_element_transactions__post
      type: object
    Body_search_element_users__post:
      properties:
        filters:
          additionalProperties: true
          default: {}
          title: Filters
          type: object
        orders:
          additionalProperties: true
          default: {}
          title: Orders
          type: object
      title: Body_search_element_users__post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PayRequest:
      properties:
        amount:
          title: Amount
          type: number
        merchant:
          title: Merchant
          type: string
      required:
      - amount
      - merchant
      title: PayRequest
      type: object
...
    SendRequest:
      properties:
        amount:
          title: Amount
          type: number
        description:
          anyOf:
          - type: string
          - type: 'null'
          default: Transfer
          title: Description
        recipient_email:
          title: Recipient Email
          type: string
      required:
      - recipient_email
      - amount
      title: SendRequest
      type: object
...
    TopUpRequest:
      properties:
        amount:
          title: Amount
          type: number
        description:
          anyOf:
          - type: string
          - type: 'null'
          default: Wallet Top-up
          title: Description
      required:
      - amount
      title: TopUpRequest
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    Wallet:
      properties:
        address:
          description: The wallet address
          title: Address
          type: string
          x_required: true
          x_unique: true
        balance:
          anyOf:
          - type: number
          - type: 'null'
          default: 0.0
          description: The wallet balance
          title: Balance
        currency:
          anyOf:
          - type: string
          - type: 'null'
          default: USD
          description: The wallet currency
          title: Currency
        id:
          default: ''
          title: Id
          type: string
        name:
          title: Name
          type: string
        owner:
          default: ''
          title: Owner
          type: string
        user_id:
          description: The user ID
          title: User Id
          type: string
          x_required: true
          x_type: objectId
      required:
      - name
      - address
      - user_id
      title: Wallet
      type: object
    WalletResponse:
      properties:
        balance:
          title: Balance
          type: number
        currency:
          title: Currency
          type: string
        transactions:
          items:
            additionalProperties: true
            type: object
          title: Transactions
          type: array
      required:
      - balance
      - currency
      - transactions
      title: WalletResponse
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: token
      type: oauth2
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /transactions/:
    post:
      operationId: search_element_transactions__post
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_search_element_transactions__post'
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Search Element Transactions  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Search Element
      tags:
      - Transaction
...
  /users/:
    post:
      operationId: search_element_users__post
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_search_element_users__post'
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Search Element Users  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Search Element
      tags:
      - User
  /wallet/:
    get:
      operationId: get_wallet_wallet__get
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Wallet
      tags:
      - Wallet
    post:
      operationId: create_element_wallet__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Wallet'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Element Wallet  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Element
      tags:
      - Wallet
  /wallet/pay:
    post:
      operationId: pay_wallet_pay_post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Pay
      tags:
      - Wallet
  /wallet/send:
    post:
      operationId: send_money_wallet_send_post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Send Money
      tags:
      - Wallet
  /wallet/shops/nearby:
    get:
      operationId: get_nearby_shops_wallet_shops_nearby_get
      parameters:
      - in: query
        name: lat
        required: true
        schema:
          title: Lat
          type: number
      - in: query
        name: lng
        required: true
        schema:
          title: Lng
          type: number
      - in: query
        name: radius
        required: false
        schema:
          default: 100
          title: Radius
          type: number
      - in: query
        name: limit
        required: false
        schema:
          default: 40
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Nearby Shops
      tags:
      - Wallet
  /wallet/topup:
    post:
      operationId: topup_wallet_topup_post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopUpRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Topup
      tags:
      - Wallet
  /wallet/transactions:
    get:
      operationId: get_transactions_wallet_transactions_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Transactions
      tags:
      - Wallet
...